Folder architeture was reorganized, 2 new tables in db and method to get token and use when calling IGDB API
