--- GameStoreAPI/Dtos/ConfirmEmail/ConfirmEmailResponseDto.cs
@@ -1,10 +1,0 @@
-﻿using Microsoft.AspNetCore.Identity;
-
-namespace GameStoreAPI.Dtos.ConfirmEmail
-{
-    public class ConfirmEmailResponseDto
-    {
-        public required string message;
-        public IEnumerable<IdentityError>? errors;
-    }
-}
