--- GameStoreAPI/Dtos/DeleteAccount/DeleteAccountResponseDto.cs
@@ -1,10 +1,0 @@
-﻿using Microsoft.AspNetCore.Identity;
-
-namespace GameStoreAPI.Dtos.DeleteAccount
-{
-    public class DeleteAccountResponseDto
-    {
-        public required string message;
-        public IEnumerable<IdentityError>? errors;
-    }
-}
