--- GameStoreAPI/Dtos/ForgotPassword/ForgotPasswordRequestDto.cs
@@ -1,11 +1,0 @@
-﻿using System.ComponentModel.DataAnnotations;
-
-namespace GameStoreAPI.Dtos.ForgotPassword
-{
-    public class ForgotPasswordRequestDto
-    {
-        [Required]
-        [EmailAddress]
-        public required string Email { get; set; }
-    }
-}
