--- GameStoreAPI/Dtos/ForgotPassword/ForgotPasswordResponseDto.cs
@@ -1,7 +1,0 @@
-﻿namespace GameStoreAPI.Dtos.ForgotPassword
-{
-    public class ForgotPasswordResponseDto
-    {
-        public required string message { get; set; }
-    }
-}
