--- GameStoreAPI/Dtos/GetAccount/GetAccountResponseDto.cs
@@ -1,8 +1,0 @@
-﻿namespace GameStoreAPI.Dtos.GetAccount
-{
-    public class GetAccountResponseDto
-    {
-        public required string? userName { get; set; }
-        public required string? email { get; set; }
-    }
-}
