--- GameStoreAPI/Dtos/LoginAccount/LoginAccountResponseDto.cs
@@ -1,9 +1,0 @@
-﻿namespace GameStoreAPI.Dtos.LoginAccount
-{
-    public class LoginAccountResponseDto
-    {
-        public string? message { get; set; }
-        public string? refreshTokenRes { get; set; }
-        public string? jwtTokenRes { get; set; }
-    }
-}
