--- GameStoreAPI/Dtos/LoginAccount/LoginAccountResquestDto.cs
@@ -1,14 +1,0 @@
-﻿using System.ComponentModel.DataAnnotations;
-
-namespace GameStoreAPI.Dtos.LoginAccount
-{
-    public class LoginAccountResquestDto
-    {
-        [Required]
-        [EmailAddress]
-        public string Email { get; init; }
-
-        [Required]
-        public string Password { get; init; }
-    }
-}
