--- GameStoreAPI/Dtos/LogoutAccount/LogoutAccountResponseDto.cs
@@ -1,7 +1,0 @@
-﻿namespace GameStoreAPI.Dtos.LogoutAccount
-{
-    public class LogoutAccountResponseDto
-    {
-        public required string message;
-    }
-}
