--- GameStoreAPI/Dtos/RefreshToken/RefreshTokenRequestDto.cs
@@ -1,10 +1,0 @@
-﻿using System.ComponentModel.DataAnnotations;
-
-namespace GameStoreAPI.Dtos.RefreshToken
-{
-    public class RefreshTokenRequestDto
-    {
-        [Required]
-        public required string refreshTokenReq;
-    }
-}
