--- GameStoreAPI/Dtos/RefreshToken/RefreshTokenResponseDto.cs
@@ -1,11 +1,0 @@
-﻿using System.ComponentModel.DataAnnotations;
-
-namespace GameStoreAPI.Dtos.RefreshToken
-{
-    public class RefreshTokenResponseDto
-    {
-        public string? message;
-        public string? refreshTokenRes;
-        public string? jwtTokenRes;
-    }
-}
