--- GameStoreAPI/Dtos/ResetPassword/ResetPasswordResponseDto.cs
@@ -1,10 +1,0 @@
-﻿using Microsoft.AspNetCore.Identity;
-
-namespace GameStoreAPI.Dtos.ResetPassword
-{
-    public class ResetPasswordResponseDto
-    {
-        public required string message;
-        public IEnumerable<IdentityError>? errors;
-    }
-}
