--- GameStoreAPI/Dtos/UpdateAccount/UpdatePasswordResponseDto.cs
@@ -1,10 +1,0 @@
-﻿using Microsoft.AspNetCore.Identity;
-
-namespace GameStoreAPI.Dtos.UpdateAccount
-{
-    public class UpdatePasswordResponseDto
-    {
-        public required string message { get; set; }
-        public IEnumerable<IdentityError>? errors { get; set; }
-    }
-}
