+++ GameStoreAPI/Features/Authentication/AuthService/EmailConfirmationStatus.cs
@@ -1,0 +1,10 @@
+﻿namespace GameStoreAPI.Features.Authentication.AuthService
+{
+    public enum EmailConfirmationStatus
+    {
+        Success,
+        UserNotFound,
+        InvalidToken,
+        Failure
+    }
+}
