+++ GameStoreAPI/Features/Authentication/Dtos/ConfirmEmail/ConfirmEmailResponseDto.cs
@@ -1,0 +1,10 @@
+﻿using Microsoft.AspNetCore.Identity;
+
+namespace GameStoreAPI.Features.Authentication.Dtos.ConfirmEmail
+{
+    public class ConfirmEmailResponseDto
+    {
+        public required string message;
+        public IEnumerable<IdentityError>? errors;
+    }
+}
