+++ GameStoreAPI/Features/Authentication/Dtos/ForgotPassword/ForgotPasswordRequestDto.cs
@@ -1,0 +1,11 @@
+﻿using System.ComponentModel.DataAnnotations;
+
+namespace GameStoreAPI.Features.Authentication.Dtos.ForgotPassword
+{
+    public class ForgotPasswordRequestDto
+    {
+        [Required]
+        [EmailAddress]
+        public required string Email { get; set; }
+    }
+}
