+++ GameStoreAPI/Features/Authentication/Dtos/ForgotPassword/ForgotPasswordResponseDto.cs
@@ -1,0 +1,7 @@
+﻿namespace GameStoreAPI.Features.Authentication.Dtos.ForgotPassword
+{
+    public class ForgotPasswordResponseDto
+    {
+        public required string message { get; set; }
+    }
+}
