+++ GameStoreAPI/Features/Authentication/Dtos/LoginAccount/LoginAccountResponseDto.cs
@@ -1,0 +1,9 @@
+﻿namespace GameStoreAPI.Features.Authentication.Dtos.LoginAccount
+{
+    public class LoginAccountResponseDto
+    {
+        public string? message { get; set; }
+        public string? refreshTokenRes { get; set; }
+        public string? jwtTokenRes { get; set; }
+    }
+}
