+++ GameStoreAPI/Features/Authentication/Dtos/LoginAccount/LoginAccountResquestDto.cs
@@ -1,0 +1,14 @@
+﻿using System.ComponentModel.DataAnnotations;
+
+namespace GameStoreAPI.Features.Authentication.Dtos.LoginAccount
+{
+    public class LoginAccountResquestDto
+    {
+        [Required]
+        [EmailAddress]
+        public string Email { get; init; }
+
+        [Required]
+        public string Password { get; init; }
+    }
+}
