+++ GameStoreAPI/Features/Authentication/Dtos/LogoutAccount/LogoutAccountResponseDto.cs
@@ -1,0 +1,7 @@
+﻿namespace GameStoreAPI.Features.Authentication.Dtos.LogoutAccount
+{
+    public class LogoutAccountResponseDto
+    {
+        public required string message;
+    }
+}
