+++ GameStoreAPI/Features/Authentication/Dtos/RefreshToken/RefreshTokenRequestDto.cs
@@ -1,0 +1,10 @@
+﻿using System.ComponentModel.DataAnnotations;
+
+namespace GameStoreAPI.Features.Authentication.Dtos.RefreshToken
+{
+    public class RefreshTokenRequestDto
+    {
+        [Required]
+        public required string refreshTokenReq;
+    }
+}
