+++ GameStoreAPI/Features/Authentication/Dtos/RefreshToken/RefreshTokenResponseDto.cs
@@ -1,0 +1,11 @@
+﻿using System.ComponentModel.DataAnnotations;
+
+namespace GameStoreAPI.Features.Authentication.Dtos.RefreshToken
+{
+    public class RefreshTokenResponseDto
+    {
+        public string? message;
+        public string? refreshTokenRes;
+        public string? jwtTokenRes;
+    }
+}
