+++ GameStoreAPI/Features/Games/Dtos/IGDBAuthResponseDto.cs
@@ -1,0 +1,16 @@
+﻿using System.Text.Json.Serialization;
+
+namespace GameStoreAPI.Features.Games.Dtos
+{
+    public class IGDBAuthResponseDto
+    {
+        [JsonPropertyName("access_token")]
+        public string AccessToken { get; set; }
+
+        [JsonPropertyName("expires_in")]
+        public int ExpiresIn { get; set; }
+
+        [JsonPropertyName("token_type")]
+        public string TokenType { get; set; }
+    }
+}
