+++ GameStoreAPI/Features/Games/GamesController.cs
@@ -1,0 +1,15 @@
+﻿using Microsoft.AspNetCore.Mvc;
+
+namespace GameStoreAPI.Features.Games
+{
+    [ApiController]
+    [Route("api/[controller]")]
+    public class GamesController : ControllerBase
+    {
+        private void getApiToken()
+        {
+            return;
+        }
+
+    }
+}
