+++ GameStoreAPI/Features/Games/GamesService/IIGDBService.cs
@@ -1,0 +1,8 @@
+﻿namespace GameStoreAPI.Features.Games.GamesService
+{
+    public interface IIGDBService
+    {
+
+        Task<string> GetValidApiToken();
+    }
+}
