+++ GameStoreAPI/Features/MyAccount/Dtos/DeleteAccount/DeleteAccountResponseDto.cs
@@ -1,0 +1,10 @@
+﻿using Microsoft.AspNetCore.Identity;
+
+namespace GameStoreAPI.Features.MyAccount.Dtos.DeleteAccount
+{
+    public class DeleteAccountResponseDto
+    {
+        public required string message;
+        public IEnumerable<IdentityError>? errors;
+    }
+}
