+++ GameStoreAPI/Features/MyAccount/Dtos/GetAccount/GetAccountResponseDto.cs
@@ -1,0 +1,8 @@
+﻿namespace GameStoreAPI.Features.MyAccount.Dtos.GetAccount
+{
+    public class GetAccountResponseDto
+    {
+        public required string? userName { get; set; }
+        public required string? email { get; set; }
+    }
+}
