--- GameStoreAPI/Services/AuthService/EmailConfirmationStatus.cs
@@ -1,10 +1,0 @@
-﻿namespace GameStoreAPI.Services.AuthService
-{
-    public enum EmailConfirmationStatus
-    {
-        Success,
-        UserNotFound,
-        InvalidToken,
-        Failure
-    }
-}
